window_setting-test5-loop entrybox
Error: The program seems to be printing on top of itself a simple fix is to make the loop of the function not start one to to avoid printing on top of the label then what's preventing the button to print on top of the entry boxs. Another fixes is to have the entry loop on A-frame.