-test2(child pointers of the node)
output"
    10
   /  \
  5    15"
-In a BST, each node can have up to two children:

Left Child: The left child node contains a value that is less than the value of the current node.
Right Child: The right child node contains a value that is greater than or equal to the value of the current node.
The "child pointers" are the attributes in a node that hold references to these child nodes. These pointers are used to traverse the tree, insert new nodes, search for values, and perform other tree operations.