-test22(ClassMain/Password/Convertor)
Successfully able to add the three classes together however there are a few improvements that need to be made First off the password they'll but it needs to be disabled or perhaps destroyed when moving on to the next class. It also should be note that the locking function on the password hasn't been properly sort it outlet yet. Next would be the calculations for the data that will be transported to the next window also the check button needs to be changed into results button and it's command should allow the user to create the results window.